+++ src/mui/layout/AppBarMobile.js
@@ -5,14 +5,38 @@ import muiThemeable from 'material-ui/styles/muiThemeable';
 import compose from 'recompose/compose';
 import { toggleSidebar as toggleSidebarAction } from '../../actions';
 
-const AppBar = ({ title, toggleSidebar }) => (
+const style = {
+    bar: {
+        height: '3em',
+        position: 'absolute',
+        top: 0,
+    },
+    title: {
+        fontSize: '1.25em',
+        lineHeight: '2.5em',
+    },
+    icon: {
+        marginTop: 0,
+        marginRight: 0,
+        marginLeft: '-24px',
+    },
+    link: {
+        color: '#fff',
+        textDecoration: 'none',
+    },
+};
+
+const AppBarMobile = ({ title, toggleSidebar }) => (
     <MuiAppBar
+        style={style.bar}
+        titleStyle={style.title}
+        iconStyleLeft={style.icon}
         title={title}
         onLeftIconButtonTouchTap={toggleSidebar}
     />
 );
 
-AppBar.propTypes = {
+AppBarMobile.propTypes = {
     title: PropTypes.oneOfType([
         PropTypes.string,
         PropTypes.element
@@ -27,4 +51,4 @@ const enhance = compose(
     }),
 );
 
-export default enhance(AppBar);
+export default enhance(AppBarMobile);