+++ src/mui/list/DatagridBody.js
@@ -1,11 +1,12 @@
 import React, { PropTypes } from 'react';
 import shouldUpdate from 'recompose/shouldUpdate';
+import { TableBody, TableRow } from 'material-ui/Table';
 import DatagridCell from './DatagridCell';
 
-const DatagridBody = ({ resource, children, ids, data, basePath, styles, rowStyle }) => (
-    <tbody style={styles.tbody}>
+const DatagridBody = ({ resource, children, ids, data, basePath, styles, rowStyle, options, rowOptions, ...rest }) => (
+    <TableBody displayRowCheckbox={false} {...rest} {...options}>
         {ids.map((id, rowIndex) => (
-            <tr style={rowStyle ? rowStyle(data[id], rowIndex) : styles.tr} key={id}>
+            <TableRow style={rowStyle ? rowStyle(data[id], rowIndex) : styles.tr} key={id} selectable={false} {...rowOptions}>
                 {React.Children.map(children, (field, index) => (
                     <DatagridCell
                         key={`${id}-${field.props.source || index}`}
@@ -14,9 +15,9 @@ const DatagridBody = ({ resource, children, ids, data, basePath, styles, rowStyl
                         {...{ field, basePath, resource }}
                     />
                 ))}
-            </tr>
+            </TableRow>
         ))}
-    </tbody>
+    </TableBody>
 );
 
 DatagridBody.propTypes = {
@@ -25,6 +26,8 @@ DatagridBody.propTypes = {
     resource: PropTypes.string,
     data: PropTypes.object.isRequired,
     basePath: PropTypes.string,
+    options: PropTypes.object,
+    rowOptions: PropTypes.object,
     styles: PropTypes.object,
     rowStyle: PropTypes.func,
 };
@@ -34,4 +37,9 @@ DatagridBody.defaultProps = {
     ids: [],
 };
 
-export default shouldUpdate((props, nextProps) => nextProps.isLoading === false)(DatagridBody);
+const PureDatagridBody = shouldUpdate((props, nextProps) => nextProps.isLoading === false)(DatagridBody);
+
+// trick material-ui Table into thinking this is one of the child type it supports
+PureDatagridBody.muiName = 'TableBody';
+
+export default PureDatagridBody;