+++ src/mui/list/index.js
@@ -4,4 +4,5 @@ export FilterButton from './FilterButton';
 export FilterForm from './FilterForm';
 export List from './List';
 export Pagination from './Pagination';
+export SimpleList from './SimpleList';
 export SingleFieldList from './SingleFieldList';