+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## v0.9.0
+
+* [BC Break] Update REST response format to always expect a `data` key
+* Add mobile UI (Sidebar, AppBar, Datagrid, actions, form action)
+* Add `<Responsive>` component
+* Add `<ViewTitle>` component (to be used instead of `<CardTitle>` for responsive views)
+* Add ability to hide sidebar using hamburger menu
+* Add `<Sidebar>` component
+* Add `menu` prop to `<Admin>`, to customize the menu without overriding the layout
+* Add dashboard MenuItem on top of Menu when relevant
+* Add ability to execute crud actions without redirect ([thedersen](https://github.com/thedersen))
+* Add support for custom `onBlur`, `onChange`, and `onFocus` props to `<NumberInput>` and `<TextInput>`
+* Add support for image preview change via dispatches form value in `<<ImageInput />`
+* Add support for custom redirect path when auth check fails ([thedersen](https://github.com/thedersen))
+* Add support for non-cancelable fetch in `crudGetList` action ([thedersen](https://github.com/thedersen))
+* Add support for default locale in `resolveBrowserLocale`
+* Add ability to translate `CRUD_UPDATE` to HTTP `PATCH` method
+* Add ability to hide fields
+* Add Spanish translations ([JonatanSalas](https://github.com/JonatanSalas))
+* Add Vietnamese translations ([kimkha](https://github.com/kimkha))
+* Improve Login page UX with form lock and loader ([Natsuke](https://github.com/Natsuke))
+* Improve `<Datagrid>` rendering options by using material-ui `<Table>` element
+* Improve loader and button components rendering speed
+* Remove link to dashboard in the top bar
+* Remove CSS copy step from build
+* Fix notification for server error even when the response body is empty
+
 ## v0.8.4
 
 * Fix `defaultValue` assignment for nested resource attributes ([cytomich](https://github.com/cytomich))